PR-E-6: extract DataStructSynthesizer (ADR-0029 synthesized members + inline struct)
Sixth PR in the ReflectionMetadataEmitter decomposition. Moves the IL
emission for ADR-0029 `data struct` synthesized members and the
`inline struct` single-field synthesized members into a new
`internal sealed` component, `DataStructSynthesizer`, in
src/Core/CodeAnalysis/Emit/DataStructSynthesizer.cs.

Methods moved (all previously top-level `private` on
ReflectionMetadataEmitter — clean Option-A extraction, no BodyEmitter
seam):

  Inline-struct group:
    EmitInlineStructSynthesizedMembers
    EmitInlineStructConstructor
    EmitInlineEqualsObject
    EmitInlineEqualsTyped
    EmitInlineGetHashCode
    EmitInlineToString
    EmitInlineEqualityOperator
    EmitInlineDeconstruct

  Data-struct group:
    EmitDataStructSynthesizedMembers
    EmitDataStructEqualsObject
    EmitDataStructEqualsTyped
    EmitDataStructGetHashCode
    EmitDataStructToString
    EmitDataStructEqualityOperator
    EmitDataStructDeconstruct

Plus three helpers exclusively used by the methods above and the
per-emit StandaloneSignatureHandle they cache:
    FinishInlineBody
    GetHashCodeCombineObjectSpec
    GetHashCodeAddObjectSpec
    GetHashCodeLocalSignature (with its hashCodeLocalSig field)

DataStructSynthesizer is constructor-injected with EmitContext,
MetadataTokenCache, WellKnownReferences, ConversionEmitter, and four
thin delegate callbacks (EncodeTypeSymbol, GetElementTypeToken,
GetTypeReference, NextParameterHandle) bound to the remaining root-
emitter helpers it depends on — same pattern PR-E-5 ConversionEmitter
established. No hard back-reference to ReflectionMetadataEmitter.

The shared DataStructOpEqualityHandles dictionary stays on
MetadataTokenCache, exposed as cache.DataStructOpEqualityHandles, so
future consumers (e.g. the operator-lookup path in the body emitter)
can keep reading it without depending on this component.

No public/internal symbol surface visible to Compilation.cs or tests
has been renamed.

LoC: ReflectionMetadataEmitter.cs 16269 → 15587 (-682).
     DataStructSynthesizer.cs: 826.
